requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pillow==10.4.0
tensorflow==2.17.0
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.8.0